requirements.txt
altgraph==0.17.3
paho-mqtt==1.6.1
pefile==2023.2.7
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
pywin32-ctypes==0.2.0
xlrd==2.0.1
